Move general components to path components
The common factual components should belong to the `components` path since they will share the permissive (BSD) license among all React apps in MathSwe, while other commercial components, like cookies will be AGPL. This is crucial to keep MathSwe Templates maintainable, too.
