+++ src/components/CloseIcon.tsx
