+++ src/ui/legal/CookieUsageTable.tsx
@@ -8,7 +8,7 @@ import {
     MathSweBaseDomain,
     MathSweDomain,
 } from "@app/legal/cookies/cookies.ts";
-import { Table } from "@ui/Table.tsx";
+import { Table } from "@components/Table.tsx";
 import { ReactNode } from "react";
 
 type CookieUsageTableProps = {
