+++ tsconfig.app.json
@@ -12,6 +12,7 @@
         "paths": {
             "@/*": [ "src/*" ],
             "@persistence/*": [ "src/persistence/*" ],
+            "@components/*": [ "src/components/*" ],
             "@ui/*": [ "src/ui/*" ],
             "@analytics/*": [ "src/analytics/*" ],
             "@app/*": [ "src/app/*" ],
